Range problem fixed
The zb3219PA has an external cc2592 attached, which needed to be setup
correctly. Found a patch on the contiki forum, which helped me solve the
issue. It might be, that I have to remove the code from the antenna init
due to standby/resume issues, but we'll see.
The output power from the cc2538 has been bumped to 5dBm