Various fixes for the node
The code is mainly from the openmote project - this meant that some
sensors were unknown to us and have been removed.
The pulsesensor has been introduced, and tested with our new relayboard.
It works ok.
The sensors are now all instantiated based on the original contiki
sensor design, but with some extras config struct added.