Implemented the mainsdetect sensor
Instead of the old GPIO interrupt with a timeout, I have implemented it
using a counter and a timeout.
After a counter has counted 5 pulses a timeout timer (105ms) is started.
For every 5 pulses the timeout timer is reset.
If pulses fails to appear, the timeout interrupt will be fired.
To avoid false triggering when mains is gone, a second timeout will be
set before retriggering the mains detect counter.